+++ src/components/App/App.tsx
@@ -1,21 +1,22 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { useConnect, useDisconnect } from "wagmi";
 import cn from "classnames";
 
 import Header from "../Header/Header";
 import { Footer } from "../Footer/Footer";
 import { Routes } from "../Routes";
-import { BugIcon } from "../../ui-kit/images/icons";
 import { ReactComponent as MetamaskIcon } from "../../ui-kit/images/metamask-circle.svg";
 import { ReactComponent as WalletConnectIcon } from "../../ui-kit/images/walletconnect.svg";
 import { ReactComponent as GnosisSafeIcon } from "../../ui-kit/images/gnosis-safe.svg";
 import { trackEvent } from "../../utils/events";
 import { useAutoConnect } from "../../hooks/useAutoConnect";
+import { StateContext } from "../../reducer/constants";
 
 import "./App.scss";
 
 const App = () => {
-    const { connect, connectors, isLoading, pendingConnector } = useConnect();
+    const { uiSelectedChainId } = useContext(StateContext);
+    const { connectAsync, connectors, isLoading, pendingConnector } = useConnect({ chainId: uiSelectedChainId });
     const { disconnect } = useDisconnect();
 
     const [isModalVisible, setIsModalVisible] = useState(false);
@@ -35,18 +36,13 @@ const App = () => {
         setIsModalVisible(false);
     };
 
-    const handleConnect = (selectedConnector: any) => () => {
+    const handleConnect = (selectedConnector: any) => async () => {
         trackEvent("connect-wallet", { name: selectedConnector.name, location: window.location.href });
-        connect({ connector: selectedConnector });
+        await connectAsync({
+            connector: selectedConnector,
+            chainId: uiSelectedChainId,
+        });
         closeModal();
-    };
-
-    const onBugClick = () => {
-        window.open(
-            // eslint-disable-next-line max-len
-            "https://github.com/StableUnit/safe-transfer/issues/new?assignees=Kud8&labels=&template=bug_report.md&title=%5BBUG%5D",
-            "_blank"
-        );
     };
 
     const renderIcon = (id: string) => {
@@ -68,7 +64,6 @@ const App = () => {
             <div className="App__container">
                 <Routes onConnect={openModal} />
             </div>
-            <BugIcon className="bug-icon" id="bug-icon" onClick={onBugClick} />
             <Footer />
             {isModalVisible && (
                 <div className="connect-modal">