+++ src/components/ReceiveForm/ReceiveForm.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { IconButton } from "@mui/material";
 import cn from "classnames";
 
-import { useAccount, useContract, useNetwork, useSigner } from "wagmi";
+import { useAccount, useContract, useContractWrite, useFeeData, useNetwork, useSigner } from "wagmi";
 import {
     NetworkType,
     getAddressLink,
@@ -34,6 +34,9 @@ import "../PageNotFound/styles.scss";
 import "./ReceiveForm.scss";
 import { useReceiveToken } from "../../hooks/useReceiveToken";
 import CONTRACT_ERC20 from "../../contracts/ERC20.json";
+import { Actions } from "../../reducer";
+import { DispatchContext } from "../../reducer/constants";
+import { useGasPrice } from "../../hooks/useGasPrice";
 
 interface TransferFormProps {
     onConnect: () => void;
@@ -45,6 +48,7 @@ const getValue = (tokenMetadata: TokenMetadataType | undefined, tokenData: Token
         : tokenData.value;
 
 const ReceiveForm = React.memo(({ onConnect }: TransferFormProps) => {
+    const dispatch = useContext(DispatchContext);
     const { data: signer } = useSigner();
     const { address } = useAccount();
     const { chain } = useNetwork();
@@ -61,11 +65,23 @@ const ReceiveForm = React.memo(({ onConnect }: TransferFormProps) => {
     }, [address]);
 
     const { tokenData, token } = useReceiveToken();
-    const tokenDataContract = useContract({
-        address: tokenData?.address,
+
+    const config = {
+        mode: "recklesslyUnprepared" as "recklesslyUnprepared" | "prepared",
+        address: tokenData?.address as `0x${string}`,
         abi: CONTRACT_ERC20,
-        signerOrProvider: signer,
-    });
+        chainId: chain?.id,
+    };
+    const tokenDataContract = useContract({ ...config, signerOrProvider: signer });
+    const { writeAsync: approve } = useContractWrite({ ...config, functionName: "approve" });
+    const { writeAsync: transferFrom } = useContractWrite({ ...config, functionName: "transferFrom" });
+    const gasPrice = useGasPrice(chain?.id);
+
+    useEffect(() => {
+        if (tokenData?.chain && dispatch) {
+            dispatch({ type: Actions.SetUISelectedChainId, payload: +networkToId[tokenData.chain] });
+        }
+    }, [tokenData?.chain, dispatch]);
 
     const [isToAddressRequesting, setIsToAddressRequesting] = useState(false);
     const [toAddress, setToAddress] = useState<string | null>();
@@ -124,8 +140,11 @@ const ReceiveForm = React.memo(({ onConnect }: TransferFormProps) => {
         setIsTransferFetching(true);
 
         try {
-            if (tokenData && tokenDataContract && networkName && toAddress) {
-                const tx = await tokenDataContract.transferFrom(tokenData.from, toAddress, tokenData.value);
+            if (tokenData && transferFrom && networkName && toAddress) {
+                const tx = await transferFrom({
+                    recklesslySetUnpreparedArgs: [tokenData.from, toAddress, tokenData.value],
+                    recklesslySetUnpreparedOverrides: { gasPrice },
+                });
                 setTrxHash(tx.hash);
                 const eventData = {
                     location: window.location.href,
@@ -160,8 +179,11 @@ const ReceiveForm = React.memo(({ onConnect }: TransferFormProps) => {
         setIsCancelFetching(true);
 
         try {
-            if (tokenData && tokenDataContract && networkName && toAddress) {
-                const tx = await tokenDataContract.approve(toAddress, "0");
+            if (tokenData && tokenDataContract && approve && networkName && toAddress) {
+                const tx = await approve({
+                    recklesslySetUnpreparedArgs: [toAddress, "0"],
+                    recklesslySetUnpreparedOverrides: { gasPrice },
+                });
                 const symbol = await tokenDataContract.symbol();
 
                 // eslint-disable-next-line max-len