+++ src/reducer/constants.ts
@@ -3,6 +3,7 @@ import { ActionType, ReducerState } from "./index";
 
 export const initialState: ReducerState = {
     newCustomToken: undefined,
+    uiSelectedChainId: 1,
 };
 
 export const StateContext = React.createContext(initialState);