+++ src/ui-kit/images/icons/index.tsx
@@ -1,3 +1,2 @@
 export { ReactComponent as BurgerIcon } from "./burger.svg";
 export { ReactComponent as CloseIcon } from "./close.svg";
-export { ReactComponent as BugIcon } from "./bug.svg";